updated ui to UI
