+++ app/src/main/java/edu/cascadia/mobas/pollinatorpathway/UI/SetUp/SetUpFragment.java
@@ -1,15 +1,12 @@
-package edu.cascadia.mobas.pollinatorpathway.ui.SetUp;
+package edu.cascadia.mobas.pollinatorpathway.UI.SetUp;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.TextView;
 
 import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
-import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProvider;
 
 import edu.cascadia.mobas.pollinatorpathway.databinding.FragmentSetupBinding;
