+++ app/src/main/res/layout/fragment_login.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="edu.cascadia.mobas.pollinatorpathway.ui.Login.LoginFragment">
+    tools:context="edu.cascadia.mobas.pollinatorpathway.UI.Login.LoginFragment">
 
 
     <TextView
