+++ app/src/main/res/layout/qrcode_fragment.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.QRcode.QRcode">
+    tools:context=".UI.QRcode.QRcode">
 
     <androidx.camera.view.PreviewView
         android:id="@+id/QRcode_previewView"
