+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -7,19 +7,19 @@
 
     <fragment
         android:id="@+id/navigation_home"
-        android:name="edu.cascadia.mobas.pollinatorpathway.ui.Home.HomeFragment"
+        android:name="edu.cascadia.mobas.pollinatorpathway.UI.Home.HomeFragment"
         android:label="@string/title_home"
         tools:layout="@layout/fragment_home" />
 
     <fragment
         android:id="@+id/navigation_setup"
-        android:name="edu.cascadia.mobas.pollinatorpathway.ui.SetUp.SetUpFragment"
+        android:name="edu.cascadia.mobas.pollinatorpathway.UI.SetUp.SetUpFragment"
         android:label="Set-Up"
         tools:layout="@layout/fragment_setup" />
 
     <fragment
         android:id="@+id/navigation_account"
-        android:name="edu.cascadia.mobas.pollinatorpathway.ui.Login.LoginFragment"
+        android:name="edu.cascadia.mobas.pollinatorpathway.UI.Login.LoginFragment"
         android:label="Login"
         tools:layout="@layout/fragment_login" />
         <action
@@ -27,12 +27,12 @@
             app:destination="@id/profileFragment" />
     <fragment
         android:id="@+id/mapsFragment2"
-        android:name="edu.cascadia.mobas.pollinatorpathway.ui.Maps.MapsFragment"
+        android:name="edu.cascadia.mobas.pollinatorpathway.UI.Maps.MapsFragment"
         android:label="Navigation Map"
         tools:layout="@layout/fragment_maps" />
     <fragment
         android:id="@+id/profileFragment"
-        android:name="edu.cascadia.mobas.pollinatorpathway.ui.Profile.ProfileFragment"
+        android:name="edu.cascadia.mobas.pollinatorpathway.UI.Profile.ProfileFragment"
         android:label="Profile"
         tools:layout="@layout/fragment_profile" />
 </navigation>
